Clean up and remove .xml layout files
