--- datacapture/src/main/res/layout/multi_select_dialog.xml
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
- Copyright 2021 Google LLC
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-      http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="@dimen/dialog_padding"
->
-
-    <TextView
-        android:id="@+id/dialog_title"
-        style="?attr/questionnaireDialogTitleStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="@dimen/dialog_title_margin_bottom"
-        tools:text="What type of head pain?"
-    />
-
-    <TextView
-        android:id="@+id/dialog_subtitle"
-        style="?attr/questionnaireSubtitleTextStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/multi_select_dialog_subtitle"
-        android:visibility="gone"
-        tools:visibility="visible"
-    />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="@dimen/dialog_option_scroll_margin_top"
-        android:scrollbars="vertical"
-        android:layout_weight="1"
-        tools:listitem="@layout/option_item_multi"
-    />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-    >
-
-        <com.google.android.material.button.MaterialButton
-            android:id="@+id/negative_button"
-            style="?attr/questionnaireDialogButtonStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/dialog_confirmation_button_padding"
-            android:text="@android:string/cancel"
-        />
-
-        <Space
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-        />
-
-        <com.google.android.material.button.MaterialButton
-            android:id="@+id/positive_button"
-            style="?attr/questionnaireDialogButtonStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/dialog_confirmation_button_padding"
-            android:text="@string/save"
-        />
-    </LinearLayout>
-
-</LinearLayout>
