--- datacapture/src/main/res/layout/option_item_other_text.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
- Copyright 2021 Google LLC
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-      http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:gravity="center_vertical"
->
-  <com.google.android.material.textfield.TextInputLayout
-        style="?attr/questionnaireTextInputLayoutStyle"
-        android:id="@+id/text_input_layout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-    >
-    <com.google.android.material.textfield.TextInputEditText
-            android:hint="@string/open_choice_other_hint"
-            style="?attr/questionnaireTextInputEditTextStyle"
-            android:id="@+id/edit_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-        />
-  </com.google.android.material.textfield.TextInputLayout>
-  <ImageView
-        android:id="@+id/delete_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_vertical"
-        android:padding="8dp"
-        android:src="@drawable/gm_remove_circle_outline_24"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        app:tint="?colorPrimary"
-    />
-</LinearLayout>
