--- datacapture/src/main/res/layout/option_item_single.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
- Copyright 2021 Google LLC
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-      http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<RadioButton
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/radio_button"
-    style="?attr/questionnaireRadioButtonStyle"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:drawablePadding="@dimen/icon_drawable_padding"
-/>
