--- datacapture/src/main/res/layout/option_select_view.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
- Copyright 2020 Google LLC
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginHorizontal="@dimen/item_margin_horizontal"
-    android:layout_marginVertical="@dimen/item_margin_vertical"
-    android:orientation="vertical"
->
-
-    <com.google.android.fhir.datacapture.views.HeaderView
-        android:id="@+id/header"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-    />
-
-    <com.google.android.fhir.datacapture.views.MediaView
-        android:id="@+id/item_media"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-    />
-
-    <com.google.android.material.textfield.TextInputLayout
-        style="?attr/questionnaireDropdownLayoutStyle"
-        android:id="@+id/multi_select_summary_holder"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-    >
-        <AutoCompleteTextView
-            android:id="@+id/multi_select_summary"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:inputType="none"
-            tools:text="Answer A, Answer B"
-        />
-    </com.google.android.material.textfield.TextInputLayout>
-</LinearLayout>
